refactor(astvisitor): AstVisitor を独立させる
