+++ src/feature/metric/AstLogger.ts
@@ -1,5 +1,5 @@
 import ts from 'typescript';
-import type { AstVisitor, VisitProps } from './AstTraverser';
+import type { AstVisitor, VisitProps } from './AstVisitor';
 
 export default class AstLogger implements AstVisitor {
   visit({ node, depth, sourceFile }: VisitProps): void {
